delete error

+++ controlers/del_prod.php
@@ -1,4 +1,5 @@
 <?php
+session_start();
 require_once "../Models/Produit.php";
 $products = new Products;
 
@@ -10,9 +11,16 @@ if (isset($_GET["id"]) and !empty($_GET["id"])) {
     $image = $_GET["image"];
 
     try {
-        unlink(PRODUCTS_FOLDER . $image);
-        $products->delete($id);
-        $res = "supprime";
+        if (!unlink(PRODUCTS_FOLDER . $image)) {
+            throw new Exception("impossible de supprimer l'image");
+        } else {
+            $products->delete($id);
+            $res = "supprime";
+            $_SESSION["socapco_alert"] = array(
+                "type" => "success",
+                "message" => "prodruit supprimer avec succès",
+            );
+        }
     } catch (EXCEPTION $e) {
         $res = "ERREUR" . $e->getMessage();
     }
